Moved Particle declaration and definition into main file
